+++ draft-packages/avatar/docs/Avatar.stories.tsx
@@ -1,9 +1,17 @@
 import * as React from "react"
 
+import { withDesign } from "storybook-addon-designs"
 import { Avatar } from "@kaizen/draft-avatar"
+import { figmaEmbed } from "../../../storybook/helpers"
 
 export default {
   title: "Avatar (React)",
+  parameters: {
+    ...figmaEmbed(
+      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=1929%3A14306"
+    ),
+  },
+  decorators: [withDesign],
 }
 
 export const InitialsPersonal = () => (