+++ draft-packages/form/docs/CheckboxField.stories.tsx
@@ -1,5 +1,7 @@
 import { CheckboxField } from "@kaizen/draft-form"
 import * as React from "react"
+import { withDesign } from "storybook-addon-designs"
+import { figmaEmbed } from "../../../storybook/helpers"
 
 type RenderProps = {
   checkedStatus: string
@@ -54,7 +56,11 @@ export default {
         import { CheckboxField } from "@kaizen/draft-form";
       `,
     },
+    ...figmaEmbed(
+      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=14462%3A196"
+    ),
   },
+  decorators: [withDesign],
 }
 
 export const InteractiveKaizenSiteDemo = () => (