+++ draft-packages/modal/docs/Modal.stories.tsx
@@ -17,8 +17,10 @@ import {
 import * as React from "react"
 import lockIcon from "@kaizen/component-library/icons/lock.icon.svg"
 import userIcon from "@kaizen/component-library/icons/user.icon.svg"
-
+import { withDesign } from "storybook-addon-designs"
 import { BenefitsSurvey, Negative } from "@kaizen/draft-illustration"
+import { figmaEmbed } from "../../../storybook/helpers"
+
 import styles from "./Modal.stories.scss"
 
 const Padding = ({
@@ -63,7 +65,11 @@ export default {
         "ModalAccessibleDescription, ModalAccessibleLabel, ModalBody, ModalFooter, " +
         'ModalHeader } from "@kaizen/draft-modal"',
     },
+    ...figmaEmbed(
+      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=1929%3A35440"
+    ),
   },
+  decorators: [withDesign],
 }
 
 export const ConfirmationPositiveKaizenSiteDemo = () => (