+++ draft-packages/form/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.1.3...@kaizen/draft-form@3.1.4) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+## [3.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.1.2...@kaizen/draft-form@3.1.3) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+## [3.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.1.1...@kaizen/draft-form@3.1.2) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+## [3.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.1.0...@kaizen/draft-form@3.1.1) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
 # [3.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.0.3...@kaizen/draft-form@3.1.0) (2021-01-28)