+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.5.7...@kaizen/draft-guidance-block@2.5.8) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.5.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.5.6...@kaizen/draft-guidance-block@2.5.7) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.5.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.5.5...@kaizen/draft-guidance-block@2.5.6) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.5.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.5.4...@kaizen/draft-guidance-block@2.5.5) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [2.5.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.5.3...@kaizen/draft-guidance-block@2.5.4) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block