+++ draft-packages/menu/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.8.2...@kaizen/draft-menu@2.8.3) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [2.8.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.8.1...@kaizen/draft-menu@2.8.2) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [2.8.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.8.0...@kaizen/draft-menu@2.8.1) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+# [2.8.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.7.1...@kaizen/draft-menu@2.8.0) (2021-02-03)
+
+
+### Features
+
+* Add active state prop to menu items ([#996](https://github.com/cultureamp/kaizen-design-system/issues/996)) ([d1e25ba](https://github.com/cultureamp/kaizen-design-system/commit/d1e25ba06b94fddffabd3e67d013cc23547e4d4f))
+
+
+
+
+
+## [2.7.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.7.0...@kaizen/draft-menu@2.7.1) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
 # [2.7.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.6.0...@kaizen/draft-menu@2.7.0) (2021-02-01)