+++ draft-packages/page-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-page-layout@1.1.2...@kaizen/draft-page-layout@1.1.3) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-page-layout
+
+
+
+
+
 ## [1.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-page-layout@1.1.1...@kaizen/draft-page-layout@1.1.2) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-page-layout