+++ legacy-packages/title-block/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@2.1.9...@kaizen/draft-title-block@2.1.10) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [2.1.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@2.1.8...@kaizen/draft-title-block@2.1.9) (2021-02-08)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [2.1.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@2.1.7...@kaizen/draft-title-block@2.1.8) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [2.1.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@2.1.6...@kaizen/draft-title-block@2.1.7) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
 ## [2.1.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@2.1.5...@kaizen/draft-title-block@2.1.6) (2021-01-19)
 
 **Note:** Version bump only for package @kaizen/draft-title-block